Ensure GlobalBindings inspector works without serialized type name
- Removed `typeName` string from `GlobalBinding` and stopped relying on `SerializedProperty` to display type.
- Updated `SceneGlobalContainerEditor` to use reflection to directly access `GlobalBinding.Type` and `GlobalBinding.Instance` at runtime.
- Marked `globalBindings` read-only in the inspector.
