+++ UnityProject/Saneject/Assets/Plugins/Saneject/Runtime/Global/GlobalBinding.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: dee2b6c937f64bf0a2ee9e76c76ae1b6
+timeCreated: 1755254951
