+++ UnityProject/Saneject/ProjectSettings/EditorBuildSettings.asset
@@ -5,15 +5,15 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Plugins/Saneject/Samples/DemoGame/StartScene.unity
     guid: 15d7f5558018bbb42a18f6ec8141f383
-  - enabled: 1
-    path: Assets/Plugins/Saneject/Samples/DemoGame/Scenes/GameScene.unity
-    guid: aa011174a9d382c4aa562d4014794a47
-  - enabled: 1
+  - enabled: 0
     path: Assets/Plugins/Saneject/Samples/DemoGame/Scenes/UIScene.unity
     guid: 51d3d79ee5eba1341a73cecefb1ef13b
+  - enabled: 0
+    path: Assets/Plugins/Saneject/Samples/DemoGame/Scenes/GameScene.unity
+    guid: aa011174a9d382c4aa562d4014794a47
   m_configObjects:
     com.unity.input.settings.actions: {fileID: -944628639613478452, guid: 052faaac586de48259a63d0c4782560b, type: 3}
   m_UseUCBPForAssetBundles: 0
